Split proxy and scheduler arguments
Keep options and flags distinct from initialization
