+++ contrib/mesos/cmd/km/kube-proxy.go
@@ -18,7 +18,8 @@ limitations under the License.
 package main
 
 import (
-	kubeproxy "k8s.io/kubernetes/cmd/kube-proxy/app"
+	"k8s.io/kubernetes/cmd/kube-proxy/app"
+	"k8s.io/kubernetes/cmd/kube-proxy/app/options"
 	"k8s.io/kubernetes/contrib/mesos/pkg/hyperkube"
 )
 
@@ -26,7 +27,7 @@ import (
 // description and flags.
 
 func NewKubeProxy() *Server {
-	config := kubeproxy.NewProxyConfig()
+	config := options.NewProxyConfig()
 
 	hks := Server{
 		SimpleUsage: hyperkube.CommandProxy,
@@ -38,13 +39,13 @@ func NewKubeProxy() *Server {
 
 	config.AddFlags(hks.Flags())
 
-	hks.Run = func(_ *Server, args []string) error {
-		s, err := kubeproxy.NewProxyServerDefault(config)
+	hks.Run = func(_ *Server, _ []string) error {
+		s, err := app.NewProxyServerDefault(config)
 		if err != nil {
 			return err
 		}
 
-		return s.Run(args)
+		return s.Run()
 	}
 
 	return &hks
